Test `validate()` against a generated builder
Rewrite `ValidatingBuilderSpec` to probe the new `ValidatingBuilder.validate()`
method through a real generated validating type instead of hand-written
`AbstractMessage`/`Builder` stubs.

The spec now uses the generated `InterestRate` message (a single `float percent`
field constrained with `(min) > 0.0`) to cover the meaningful paths: valid
content yields no violations, invalid content returns the violations without
throwing, and probing leaves the builder content intact.

Because a generated `ValidatingBuilder` always builds a `ValidatableMessage`,
real generated types cannot reach the "message does not support validation"
fallback branch; that branch is defensive and no longer has a dedicated stub
test.

The spec moves to the `tests:validating` module: `jvm-runtime` depends on
`spine-base` and does not apply the `io.spine.validation` code generator, so it
cannot produce a generated `ValidatableMessage`/`ValidatingBuilder` to test
against — only a plugin-applied module such as `tests:validating` can.
